Weapon speed fixed, Projectiles fixed

+++ src/Weapon.java
@@ -1,25 +1,25 @@
 public abstract class Weapon {
 
     private int xPosition;
-    private int yPosition;
-    private int speed;
+    private float yPosition;
+    private float speed;
 
 
     public int getXPosition(){
         return xPosition;
     }
-    public int getYPosition(){
+    public float getYPosition(){
         return yPosition;
     }
     public int setXPosition(int xPosition){
         this.xPosition = xPosition;
         return xPosition;
     }
-    public int setYPosition(int yPosition){
+    public float setYPosition(float yPosition){
         this.yPosition = yPosition;
         return yPosition;
     }
-    public void setSpeed(int speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
 }
